src/api/requirements.txt
fastapi
mangum
pydantic
slowapi
python-jose[cryptography]
httpx
